Refactor per review (cycle 1)
Step 12.5 review surfaced 5 findings — all addressed.

- F1 [cli/lab.py:1-490]: file was 490 LOC, ~310 of which were
  continuity-audit-only (helpers + command). Split: extract every
  continuity helper and the command itself to a new
  src/aptl/cli/continuity.py (350 LOC). lab.py drops to 170 LOC
  focused on lifecycle. Command stays at `aptl lab continuity-audit`
  via `app.command("continuity-audit")(continuity_audit)` — no UX
  change.

- F2 [cli/lab.py:172-183]: `_resolve_run_store` was a near-clone of
  `cli/runs.py:_get_store`. Hoist a single `resolve_run_store(project_dir,
  config=None)` into cli/_common.py. Both callers now thin-alias
  through it. cli/runs.py keeps the same `_get_store` name to avoid
  churning unrelated tests.

- F3 [cli/lab.py:371-491]: `continuity_audit` body was ~120 LOC,
  mixing config resolution, target validation, whitelist policy,
  archive resolution, audit dispatch, presentation, and exit-code
  policy. Two pure helpers extracted in the new module:
  `_load_kali_ips_or_exit(wl_path) -> set[str]` (whitelist policy)
  and `_continuity_exit_code(result) -> int` (exit-code policy).
  Command body now reads as a clean linear pipeline well under
  the 80-LOC threshold.

- F4 [core/continuity.py:66-73]: ContinuityAuditError was a bare
  Exception subclass; the project's exception base for
  scenario-runtime concerns is `aptl.core.scenarios.ScenarioError`.
  Rebased onto ScenarioError so a single `except ScenarioError` in
  higher-level orchestration catches both this and ScenarioStateError
  without two except arms.

- F5 [core/continuity.py:557-645]: `audit_and_revert` body was 88
  LOC, interleaving the per-target loop and the archive-emission
  block. Extract `_persist_events(run_store, run_id, events) -> str | None`
  for the persistence path. Orchestration loop is now ~20 lines
  and the archive-failure branch is independently testable.

Test patches updated to target `aptl.cli.continuity.<name>` instead
of source modules, since the new module imports them at top level
(eager) rather than inside the function body (lazy). All 108 unit +
CLI tests passing.
